orgaise folders 1

+++ depth_estimation/detect_depth2.py
@@ -5,8 +5,6 @@ import torchvision.transforms as transforms
 from torchvision import models
 from PIL import Image
 
-
-# Load YOLO model
 yolo_net = cv2.dnn.readNet("C:\Rohit\Projects\SixthSense\model\yolov3.weights", "C:\Rohit\Projects\SixthSense\model\yolov3.cfg")
 layer_names = yolo_net.getLayerNames()
 output_layers = [layer_names[i-1] for i in yolo_net.getUnconnectedOutLayers()]
@@ -14,19 +12,16 @@ output_layers = [layer_names[i-1] for i in yolo_net.getUnconnectedOutLayers()]
 with open("C:\Rohit\Projects\SixthSense\model\coco.names", "r") as f:
     classes = [line.strip() for line in f.readlines()]
 
-# Load MiDaS Depth Estimation Model
-midas = torch.hub.load("intel-isl/MiDaS", "MiDaS_small")  # Lightweight model
+midas = torch.hub.load("intel-isl/MiDaS", "MiDaS_small")  
 midas.eval()
 
-# Define MiDaS Image Preprocessing
 midas_transform = transforms.Compose([
-    transforms.Resize((256, 256)),  # Resize input
-    transforms.ToTensor(),  # Convert to tensor
-    transforms.Normalize(mean=[0.5], std=[0.5])  # Normalize
+    transforms.Resize((256, 256)),  
+    transforms.ToTensor(),  
+    transforms.Normalize(mean=[0.5], std=[0.5])  
 ])
 
 mobile_camera_url = "http://192.168.29.67:8080/video"
-# Capture Video
 cap = cv2.VideoCapture(mobile_camera_url)
 
 while True:
@@ -36,19 +31,17 @@ while True:
 
     height, width, _ = frame.shape
 
-    midas_input = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)  # Convert to RGB
-    midas_input = Image.fromarray(midas_input)  # Convert to PIL Image
-    midas_input = midas_transform(midas_input).unsqueeze(0)  # Apply transform
-    # Perform Depth Estimation
+    midas_input = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)  
+    midas_input = Image.fromarray(midas_input)  
+    midas_input = midas_transform(midas_input).unsqueeze(0)  
     with torch.no_grad():
         depth_map = midas(midas_input)
     
     depth_map = depth_map.squeeze().cpu().numpy()
     depth_map = cv2.resize(depth_map, (width, height))
-    depth_map = (depth_map - depth_map.min()) / (depth_map.max() - depth_map.min())  # Normalize
-    depth_map = depth_map.astype(np.float32)  # Ensure depth_map is float32
+    depth_map = (depth_map - depth_map.min()) / (depth_map.max() - depth_map.min())  
+    depth_map = depth_map.astype(np.float32)  
 
-    # Object Detection with YOLO
     blob = cv2.dnn.blobFromImage(frame, 0.00392, (416, 416), (0, 0, 0), True, crop=False)
     yolo_net.setInput(blob)
     outs = yolo_net.forward(output_layers)
@@ -85,15 +78,12 @@ while True:
             confidence = confidences[i]
             color = (0, 255, 0)
 
-            # Estimate Distance Using Depth Map
-            object_depth = np.mean(depth_map[y:y+h, x:x+w])  # Average depth in bounding box
+            object_depth = np.mean(depth_map[y:y+h, x:x+w])  
 
-            # Draw Bounding Box and Depth
             cv2.rectangle(frame, (x, y), (x + w, y + h), color, 2)
             cv2.putText(frame, f"{label} {confidence:.2f}", (x, y - 10), cv2.FONT_HERSHEY_SIMPLEX, 0.5, color, 2)
             cv2.putText(frame, f"Depth: {object_depth:.2f} (normalized)", (x, y + h + 20), cv2.FONT_HERSHEY_SIMPLEX, 0.5, color, 2)
 
-    # Show Results
     cv2.imshow("YOLO + AI Depth Estimation", frame)
 
     if cv2.waitKey(1) & 0xFF == ord('q'):
